partner_chains_cardano_offchain

Module assemble_and_submit_tx

Code for adding signatures to multisig transaction and submitting it.

Functions

assemble_and_submit_tx

Adds witnesses to transaction and submits it with ogmios_client.