partner_chains_cardano_offchain

Module bridge

Deposits to (in the future withdrawals from as well) bridge validator Offchain actions to initialize bridge (IlliquidCirculionSupply) and make deposits.

Functions

create_validator_utxos

Creates “blessed” UTXOs at the ICS (Bridge) validator.

deposit_with_ics_spend

This function deposits bridge token to the Illiquid Circulation Supply from the payment wallet. It does spend existing UTXO (containing ‘auth token’) at the validator address.

deposit_without_ics_input

This function deposits bridge token to the Illiquid Circulation Supply from the payment wallet. It does not consume existing UTXO at the validator address.

init_ics_scripts

Stores smart contracts used for bridge (Illiquid Circulation Supply) in the versioning system. Scripts stored are: