partner_chains_cardano_offchain::bridge

Function deposit_without_ics_input

pub async fn deposit_without_ics_input<C: QueryLedgerState + QueryNetwork + Transactions + QueryUtxoByUtxoId, A: AwaitTx>(
    genesis_utxo: UtxoId,
    token: AssetId,
    amount: NonZero<u64>,
    pc_address: &[u8],
    payment_signing_key: &CardanoPaymentSigningKey,
    client: &C,
    await_tx: &A,
) -> Result<McTxHash>

This function deposits bridge token to the Illiquid Circulation Supply from the payment wallet. It does not consume existing UTXO at the validator address.

genesis_utxo: UTxO identifying the Partner Chain.

amount number of tokens to be deposited.

pc_address information to partner chain node, where to transfer funds

payment_signing_key: Signing key of the party paying fees.

await_tx: AwaitTx strategy.