partner_chains_cardano_offchain

Module permissioned_candidates

Supports Permissioned Candidates upsert

Permissioned candidates are stored on chain in an UTXO at the Permissioned Candidates Validator address. There should be at most one UTXO at the validator address and it should contain the permissioned candidates list. This UTXO should have 1 token of the Permissioned Candidates Policy with an empty asset name. The datum encodes Permissioned Candidates using VersionedGenericDatum envelope with the Permissioned Candidates stored in the datum field of it. Field should contain list of list, where each inner list is a triple of byte strings [sidechain_public_key, aura_public_key, grandpa_publicKey].

Functions

get_permissioned_candidates

Returns all permissioned candidates.

upsert_permissioned_candidates

Upserts permissioned candidates list. Arguments: