partner_chains_cardano_offchain

Module scripts_data

Provides synthesized scripts data

Structs

Addresses

Bech32 addresses of applied validators in partner-chains smart contracts.

PlutusScriptData

Type representing a PlutusScript validator and policy pair.

PolicyIds

Policy IDs of applied scripts in partner-chains smart contracts.

ScriptsData

Provides convenient access to the addresses and hashes of the partner chain smart contracts. Data in this struct is derived from the smart contracts, applied parameters and the network.

Functions

get_scripts_data

Returns ScriptsData of the smart contracts for the partner chain identified by genesis_utxo.

get_scripts_data_with_ogmios

Returns ScriptsData of the smart contracts for the partner chain identified by genesis_utxo, for the network configured in client.

version_oracle

Returns version oracle data required by other scripts.