partner_chains_cardano_offchain

Module update_governance

Supports governance updates Building and submitting of the transaction that changes the current governance authority

The transaction:

Burns the previous governance UTXO from the version oracle validator address

Mints exactly 1 multi-sig policy token as authentication

Produces a new governance UTXO at the version oracle validator address with a version oracle Plutus datum attached that contains the script ID (32) and policy hash.

Functions

run_update_governance

Updates governance parameters.