Crate partner_chains_cardano_offchain

Off-chain code for Partner Chains Smart Contracts and general purpose utilities related to Cardano

Modules

assemble_tx

await_tx

Primitives used for awaiting for tx being observed on the blockchain

cardano_keys

Parsing and wrapping of Cardano keys

csl

General purpose code for interacting with cardano-serialization-lib

d_param

Supports D-Parameter upsert D-parameter is stored on chain in an UTXO at the D-parameter validator address. There should be at most one UTXO at the validator address and it should contain the D-parameter. This UTXO should have 1 token of the D-parameter policy with an empty asset name. The datum encodes D-parameter using VersionedGenericDatum envelope with the D-parameter being datum field being [num_permissioned_candidates, num_registered_candidates].

governance

Governance data types

init_governance

Supports governance initialization

permissioned_candidates

Supports Permissioned Candidates upsert

plutus_script

Utilities for handling Plutus script data

register

Supports candidate registration

reserve

All smart-contracts related to Rewards Token Reserve Management

scripts_data

Provides synthetized scripts data

update_governance

Supports governance updates Building and submitting of the transaction that changes the current governance authority

Enums

OffchainError