Function partner_chains_cardano_offchain::reserve::handover::handover_reserve

pub async fn handover_reserve<T: QueryLedgerState + Transactions + QueryNetwork + QueryUtxoByUtxoId, A: AwaitTx>(
    genesis_utxo: UtxoId,
    payment_key: &CardanoPaymentSigningKey,
    client: &T,
    await_tx: &A,
) -> Result<McTxHash>

Spends current UTXO at validator address to illiquid supply validator and burn reserve auth policy token, preventing further operations.