partner_chains_cardano_offchain::reserve

Module init

Initialization of the reserve management is execution of three similar transaction to initialize three scripts: Rerserve Management Validator, Reserve Management Policy and Illiquid Circulation Supply Validator.

Transaction for each of these scripts should have:

an output to Version Oracle Validator address that should:

have script reference with the script being initialized attached, script should be applied with Version Oracle Policy Id

contain 1 token of Version Oracle Policy with “Version oracle” asset name, minted in this transaction

mint redeemer should be Constr(1, [Int: SCRIPT_ID, Bytes: Applied Script Hash])

have Plutus Data that is [Int: SCRIPT_ID, Bytes: Version Oracle Policy Id]

an output to the current governance (holder of governance token) that should:

contain a new Goveranance Policy token, minted in this transaction,

mint redeemer should be empty contructor Plutus Data

a script reference input of the current Goveranance UTXO

signature of the current goveranance

Functions

init_reserve_management